// src/types/adopt.ts
export interface RawAnimalRecord {
  animal_id: number;
  animal_subid: string;
  animal_area_pkid: number;
  animal_shelter_pkid: number;
  animal_place: string;
  animal_kind: string;
  animal_Variety: string;
  animal_sex: string;
  animal_bodytype: string;
  animal_colour: string;
  animal_age: string;
  animal_sterilization: string;
  animal_bacterin: string;
  animal_foundplace: string;
  animal_title: string;
  animal_status: string;
  animal_remark: string;
  animal_caption: string;
  animal_opendate: string;
  animal_closeddate: string;
  animal_update: string;
  animal_createtime: string;
  shelter_name: string;
  album_file: string;
  album_update: string;
  cDate: string;
  shelter_address: string;
  shelter_tel: string;
}

export interface RawShelterRecord {
  ID: string;
  ShelterName: string;
  CityName: string;
  Address: string;
  Phone: string;
  OpenTime: string;
  Url: string;
  Lon: string;
  Lat: string;
  Seq: number;
}